expenses: CurrencySelect, moved
