+++ src/components/expenses/Form.tsx
@@ -23,7 +23,7 @@ import LinkButton from 'components/common/LinkButton'
 import VaultSelect from './VaultSelect'
 import ExpenseTypeSelect from './ExpenseTypeSelect'
 import ExpenseStatusSelect from './ExpenseStatusSelect'
-import CurrencySelect from './CurrencySelect'
+import CurrencySelect from '../currency/CurrencySelect'
 import DocumentSelect from './DocumentSelect'
 import PersonSelect from './PersonSelect'
 import DeletedCheckbox from './DeletedCheckbox'
